@@ -1,5 +1,6 @@
 import classNames from 'classnames';
 import * as React from 'react';
+import { useIntl } from 'react-intl';
 
 export type CellData = string | number | React.ReactNode | undefined;
 
@@ -23,7 +24,7 @@ export interface TableState {
     selectedRowKey?: string;
 }
 
-interface TableProps {
+export interface TableProps {
     /**
      * Data which is used to render Table. The first element
      * of array is used to render table head unless `noHead`
@@ -78,150 +79,81 @@ interface TableProps {
 /**
  * Cryptobase Table overrides default table
  */
-class Table extends React.Component<TableProps, TableState> {
-    constructor(props: TableProps) {
-        super(props);
+export const Table: React.FC<TableProps> = (props: TableProps) => {
+    const { formatMessage } = useIntl();
+    const [activeFilter, setActiveFilter] = React.useState<TableState['activeFilter']>(undefined);
+    const [resultData, setResultData] = React.useState<TableState['resultData']>(undefined);
+    const [selectedRowKey, setSelectedRowKey] = React.useState<TableState['selectedRowKey']>(props.selectedKey);
 
-        this.state = {
-            activeFilter: undefined,
-            resultData: undefined,
-            selectedRowKey: props.selectedKey,
-        };
-    }
+    const {
+        data,
+        header,
+        titleComponent,
+        filters = [],
+        rowKeyIndex,
+        onSelect,
+        rowBackground,
+        side,
+        rowBackgroundColor = 'rgba(184, 233, 245, 0.7)',
+    } = props;
 
-    public componentDidMount() {
-        const { filters } = this.props;
-        if (filters && filters.length > 0) {
-            this.handleFilter(filters[0]);
-        }
-    }
+    const cn = React.useMemo(() => classNames('cr-table-header__content', {
+        'cr-table-header__content-empty': !titleComponent && filters.length === 0,
+    }), [titleComponent, filters.length]);
 
-    public componentWillReceiveProps(next: TableProps) {
-        if (this.state.selectedRowKey !== next.selectedKey) {
-            this.setState({selectedRowKey: next.selectedKey});
-        }
-    }
-
-    public componentDidUpdate(prevProps: TableProps) {
-        if (prevProps.data !== this.props.data && this.props.filters) {
-            const activeFilter = this.props.filters.find(
-                filter => filter.name === this.state.activeFilter,
-            );
-
-            if (activeFilter) {
-                this.handleFilter(activeFilter);
-            }
-        }
-    }
-
-    public render() {
-        const { data, header, titleComponent, filters = [], rowKeyIndex } = this.props;
-
-        this.ensureDataIsValid(data);
-
-        const cn = classNames('cr-table-header__content', {
-            'cr-table-header__content-empty': !titleComponent && filters.length === 0,
-        });
-
-        return (
-            <div className="cr-table-container">
-                <div className={cn}>
-                    {titleComponent ? this.renderTitleComponent() : null}
-                    {filters.length
-                        ?
-                        <div className="cr-table__filters">{this.renderFilters()}</div>
-                        : null}
-                </div>
-                <table className={'cr-table'}>
-                    {header && header.length && this.renderHead(header)}
-                    {this.renderBody(data, rowKeyIndex)}
-                </table>
-                {this.renderBackground(data)}
-            </div>
-        );
-    }
-
-    private renderTitleComponent() {
-        const { titleComponent } = this.props;
-
-        return <div className={'cr-title-component'}>{titleComponent}</div>;
-    }
-
-    private renderRowCells(row: CellData[]) {
-
+    const renderRowCells = React.useCallback((row: CellData[]) => {
         return row && row.length ?
             row.map((c, index: number) =>
-                <td key={index} colSpan={row.length === 1 ? this.props.colSpan : undefined}>{c}</td>) : [];
-    }
+                <td key={index} colSpan={row.length === 1 ? props.colSpan : undefined}>{c}</td>)
+            : <td className="cr-table--no-data" colSpan={header && header.length}>{formatMessage({ id: 'page.noDataToShow' })}</td>;
+    }, [header, props.colSpan, formatMessage]);
 
-    private handleFilter(item: Filter) {
-        const { data } = this.props;
-
+    const handleFilter = React.useCallback((item: Filter) => {
         if (!item.filter) {
-            this.setState({
-                resultData: data,
-            });
+            setResultData(props.data);
 
             return;
         }
-        const resultData: CellData[][] = [...data].filter(item.filter);
-        this.setState({
-            activeFilter: item.name,
-            resultData: resultData,
-        });
-    }
+        setActiveFilter(item.name);
+        setResultData([...data].filter(item.filter));
+    }, [data, props.data]);
 
-    private handleSelect = (key: string) => () => {
-        const { onSelect } = this.props;
-
+    const handleSelect = React.useCallback((key: string) => () => {
         if (onSelect) {
-            this.setState({
-                selectedRowKey: key,
-            }, () => {
-                if (onSelect) {
-                    onSelect(key);
-                }
-            });
+            setSelectedRowKey(key);
+
+            if (onSelect) {
+                onSelect(key);
+            }
         }
-    };
+    }, [onSelect]);
 
-    private renderFilters() {
-        const { filters = [] } = this.props;
-        const { activeFilter } = this.state;
-
-        const cn = (filterName: string) => classNames('cr-table__filter', {
+    const renderFilters = React.useCallback(() => {
+        const getClassName = (filterName: string) => classNames('cr-table__filter', {
             'cr-table__filter--active': activeFilter === filterName,
         });
 
-        return filters.map((item: Filter) => {
-            const handleFilterClick = () => {
-                this.handleFilter(item);
-            };
+        return filters.map((item: Filter) => (
+            <div
+                className={getClassName(item.name)}
+                key={item.name}
+                onClick={() => handleFilter(item)}
+            >
+                {item.name}
+            </div>
+        ));
+    }, [activeFilter, filters, handleFilter]);
 
-            return (
-                <div
-                    className={cn(item.name)}
-                    key={item.name}
-                    onClick={handleFilterClick}
-                >
-                    {item.name}
-                </div>
-            );
-        });
-    }
-
-    private renderHead(row: CellData[]) {
+    const renderHead = React.useCallback((row: CellData[]) => {
         const cells = row.map((c, index) => c ?  <th key={index}>{c}</th> : <th key={index}>&nbsp;</th>);
-
         return (
             <thead className={'cr-table__head'}>
-                <tr className={'cr-table__head-row'}>{cells}</tr>
+            <tr className={'cr-table__head-row'}>{cells}</tr>
             </thead>
         );
-    }
+    }, []);
 
-    private renderRowBackground(i: number) {
-        const { rowBackground, rowBackgroundColor = 'rgba(184, 233, 245, 0.7)' } = this.props;
+    const renderRowBackground = React.useCallback((i: number) => {
         const rowBackgroundResult = rowBackground ? rowBackground(i) : {};
         const style = {
             ...rowBackgroundResult,
@@ -231,13 +163,11 @@ class Table extends React.Component<TableProps, TableState> {
         return (rowBackground
             ? <span key={i} style={style} className="cr-table-background__row" />
             : null);
-    }
+    }, [rowBackground, rowBackgroundColor]);
 
-    private renderBackground(rows: CellData[][]) {
-        const { resultData } = this.state;
-        const { rowBackground, side } = this.props;
+    const renderBackground = React.useCallback((rows: CellData[][]) => {
         const dataToBeMapped = resultData || rows;
-        const renderBackgroundRow = (r: CellData[], i: number) => this.renderRowBackground(i);
+        const renderBackgroundRow = (r: CellData[], i: number) => renderRowBackground(i);
 
         const className = classNames('cr-table-background', {
             'cr-table-background--left': side === 'left',
@@ -249,26 +179,24 @@ class Table extends React.Component<TableProps, TableState> {
                 {rowBackground && dataToBeMapped.map(renderBackgroundRow)}
             </div>
         );
-    }
+    }, [resultData, side, renderRowBackground, rowBackground]);
 
-    private renderBody(rows: CellData[][], rowKeyIndex: number | undefined) {
-        const { resultData, selectedRowKey } = this.state;
-
+    const renderBody = React.useCallback((rows: CellData[][], rowKeyIndexValue: number | undefined) => {
         const rowClassName = (key: string) => classNames({
             'cr-table__row--selected': selectedRowKey === key,
         });
 
         const dataToBeMapped = resultData || rows;
         const rowElements = dataToBeMapped.map((r, i) => {
-            const rowKey = String((rowKeyIndex !== undefined) ? r[rowKeyIndex] : i);
+            const rowKey = String((rowKeyIndexValue !== undefined) ? r[rowKeyIndexValue] : i);
 
             return (
                 <tr
                     className={rowClassName(rowKey)}
                     key={rowKey}
-                    onClick={this.handleSelect(rowKey)}
+                    onClick={handleSelect(rowKey)}
                 >
-                    {this.renderRowCells(r)}
+                    {renderRowCells(r)}
                 </tr>
             );
         });
@@ -278,19 +206,39 @@ class Table extends React.Component<TableProps, TableState> {
             {rowElements}
             </tbody>
         );
-    }
+    }, [handleSelect, renderRowCells, resultData, selectedRowKey]);
 
-    private ensureDataIsValid(data: CellData[][]) {
-        const length = data[0].length;
-        const len = data.length;
-        for (let i = 0; i < len; i += 1) {
-            if (data[i].length !== length) {
-                throw Error('Array elements must have the same length');
+    React.useEffect(() => {
+        if (props.filters) {
+            const newActiveFilter = props.filters.find(
+                filter => filter.name === activeFilter,
+            );
+
+            if (newActiveFilter) {
+                handleFilter(newActiveFilter);
             }
         }
-    }
-}
+    });
 
-export {
-    Table,
+    React.useEffect(() => {
+        setSelectedRowKey(props.selectedKey);
+    }, [props.selectedKey]);
+
+
+    return (
+        <div className="cr-table-container">
+            <div className={cn}>
+                {titleComponent ? <div className={'cr-title-component'}>{props.titleComponent}</div> : null}
+                {filters.length
+                    ?
+                    <div className="cr-table__filters">{renderFilters()}</div>
+                    : null}
+            </div>
+            <table className={'cr-table'}>
+                {header && header.length && renderHead(header)}
+                {renderBody(data, rowKeyIndex)}
+            </table>
+            {renderBackground(data)}
+        </div>
+    );
 };
