@@ -1,6 +1,9 @@
 import classnames from 'classnames';
 import * as React from 'react';
-import { CellData, FilterInput, Table } from '../';
+import { DEFAULT_MARKET_HEADERS } from '../../constants';
+import { hasDuplicates } from '../../helpers';
+import { FilterInput } from '../FilterInput';
+import { CellData, Table } from '../Table';
 
 export interface MarketsProps {
     /**
@@ -38,81 +41,21 @@ export interface MarketsProps {
     filterPlaceholder?: string;
 }
 
-export interface MarketsState {
-    /**
-     * Keeps filtered data
-     */
-    filteredData: CellData[][];
-    /**
-     * Keeps search key
-     */
-    searchKey: string;
-}
+export const Markets = (props: MarketsProps) => {
+    const [searchKey, setSearchKey] = React.useState('');
 
-export class Markets extends React.Component<MarketsProps, MarketsState> {
-    constructor(props: MarketsProps) {
-        super(props);
+    const { headers, title, filterPlaceholder = '', rowKeyIndex, selectedKey, data, filters } = props;
 
-        this.state = {
-            filteredData: props.data,
-            searchKey: '',
-        };
-    }
+    const searchFilter = React.useCallback(
+        (row: CellData[], key: string) => {
+            setSearchKey(key);
 
-    private defaultHeaders: string[] = ['Pair', 'Price', '24h Change'];
+            return (row[0] as string).toLowerCase().includes(searchKey.toLowerCase());
+        },
+        [searchKey]
+    );
 
-    public componentWillReceiveProps(nextProps: MarketsProps) {
-        if (nextProps.data !== this.props.data) {
-            this.setState({
-                filteredData: nextProps.data.filter(w => (w[0] as string).toLowerCase().includes(this.state.searchKey.toLowerCase())),
-            });
-        }
-    }
-
-    public render() {
-        const { filteredData } = this.state;
-        const { filters = true, headers, title, filterPlaceholder = '', rowKeyIndex, selectedKey } = this.props;
-        let tableData = filteredData.length > 0
-            ? filteredData
-            : [['', '', '']];
-        tableData = tableData.map(row => row.map(this.mapRows));
-
-        return (
-            <div className="cr-markets">
-                <Table
-                    data={tableData}
-                    rowKeyIndex={rowKeyIndex}
-                    selectedKey={selectedKey}
-                    filters={filters ? this.filters : []}
-                    header={headers ? headers : this.defaultHeaders}
-                    onSelect={this.props.onSelect}
-                    titleComponent={title ? title : 'Markets'}
-                />
-                <FilterInput
-                    data={this.props.data}
-                    onFilter={this.handleFilter}
-                    filter={this.searchFilter}
-                    placeholder={filterPlaceholder}
-                />
-            </div>
-        );
-    }
-
-    public searchFilter = (row: CellData[], searchKey: string) => {
-        this.setState({
-            searchKey: searchKey,
-        });
-
-        return (row[0] as string).toLowerCase().includes(searchKey.toLowerCase());
-    };
-
-    public handleFilter = (result: object[]) => {
-        this.setState({
-            filteredData: [...result] as CellData[][],
-        });
-    };
-
-    private renderChange(cell: string) {
+    const renderChange = React.useCallback((cell: string) => {
         const isItChangeValue = (c: string) => {
             return c.search('\\+') ? 'negative' : 'positive';
         };
@@ -123,52 +66,79 @@ export class Markets extends React.Component<MarketsProps, MarketsState> {
         });
 
         return <span className={className}>{cell}</span>;
-    }
+    }, []);
 
-    private mapRows = (cell: CellData) => {
-        const isChangeValue = typeof(cell) === 'string' && (cell.charAt(0) === '-' || cell.charAt(0) === '+');
+    const mapRows = React.useCallback(
+        (cell: CellData) => {
+            const isChangeValue = typeof cell === 'string' && (cell.charAt(0) === '-' || cell.charAt(0) === '+');
 
-        return isChangeValue ? this.renderChange(cell as string) : cell;
-    };
+            return isChangeValue ? renderChange(cell as string) : cell;
+        },
+        [renderChange]
+    );
 
-    private filterType = (headerKey: string, searchKey: string) => (item: CellData[]) => {
-        const typeIndex = this.props.headers ? this.props.headers.indexOf(headerKey) : this.defaultHeaders.indexOf(headerKey);
+    const filterType = React.useCallback(
+        (headerKey: string, key: string) => (item: CellData[]) => {
+            const typeIndex = (props.headers || DEFAULT_MARKET_HEADERS).indexOf(headerKey);
 
-        return (item[typeIndex] as string).includes(searchKey);
-    };
+            return (item[typeIndex] as string).includes(key);
+        },
+        [props.headers]
+    );
 
-    private get filters() {
-        const { data } = this.props;
-
-        const currencyFilters = data && data.length > 0
-            ? this.props.data
-                .map(this.getMarketFromDataRow)
-                .reduce(this.createUniqueCurrencies, [])
-                .map(this.transformCurrencyToFilter)
-            : [];
-
-        return [
-            {
-                name: 'All',
-                filter: this.filterType('Pair', ''),
-            },
-            ...currencyFilters,
-        ];
-    }
-
-    private getMarketFromDataRow = (market: React.ReactNode[]) => market[0] as string;
-
-    private createUniqueCurrencies(currencies: string[], market: string) {
-        const isCurrencyUnique = (currency: string) => !currencies.includes(currency);
-
+    const createUniqueCurrencies = React.useCallback((currencies: string[], market: string) => {
         const marketCurrencies = market.split('/').map((c: string) => c.trim());
-        const uniqueCurrencies = marketCurrencies.filter(isCurrencyUnique);
+        const uniqueCurrencies = marketCurrencies.filter((c) => !hasDuplicates(currencies, c));
 
         return currencies.concat(uniqueCurrencies);
-    }
+    }, []);
 
-    private transformCurrencyToFilter = (currency: string) => ({
-        name: currency,
-        filter: this.filterType('Pair', currency),
-    });
-}
+    const transformCurrencyToFilter = React.useCallback(
+        (currency: string) => ({
+            name: currency,
+            filter: filterType('Pair', currency),
+        }),
+        [filterType]
+    );
+
+    const getFilters = React.useCallback(() => {
+        const currencyFilters =
+            data && data.length > 0
+                ? data
+                      .map((market: React.ReactNode[]) => market[0] as string)
+                      .reduce(createUniqueCurrencies, [])
+                      .map(transformCurrencyToFilter)
+                : [];
+
+        return filters
+            ? [
+                  {
+                      name: 'All',
+                      filter: filterType('Pair', ''),
+                  },
+                  ...currencyFilters,
+              ]
+            : [];
+    }, [createUniqueCurrencies, filterType, transformCurrencyToFilter, data, filters]);
+
+    const getTableData = React.useMemo(() => {
+        const fd = data.filter(w => (w[0] as string).toLowerCase().includes(searchKey.toLowerCase()));
+
+        return fd.map((row) => row.map(mapRows));
+    }, [data, mapRows, searchKey]);
+
+    return (
+        <div className="cr-markets">
+            <Table
+                data={getTableData}
+                rowKeyIndex={rowKeyIndex}
+                selectedKey={selectedKey}
+                filters={getFilters()}
+                header={headers || DEFAULT_MARKET_HEADERS}
+                onSelect={props.onSelect}
+                titleComponent={title || 'Markets'}
+            />
+            <FilterInput data={props.data} filter={searchFilter} placeholder={filterPlaceholder} />
+        </div>
+    );
+};
