@@ -1,8 +1,9 @@
-import { mount, shallow, ShallowWrapper } from 'enzyme';
+import { shallow, ShallowWrapper } from 'enzyme';
+import { TestComponentWrapper } from 'lib/test';
 import * as React from 'react';
-import { SinonSpy, spy } from 'sinon';
+import { spy } from 'sinon';
 import { CellData } from '../../components/Table';
-import {  mapValues, OrderBook, OrderBookProps } from './';
+import { mapValues, OrderBook, OrderBookProps } from './';
 
 const data: CellData[][] = [
     ['12349', '14', '12'],
@@ -20,7 +21,11 @@ const defaultProps: OrderBookProps = {
 };
 
 const setup = (props: Partial<OrderBook> = {}) =>
-    shallow(<OrderBook {...{ ...defaultProps, ...props }} />);
+    shallow(
+        <TestComponentWrapper>
+            <OrderBook {...{ ...defaultProps, ...props }} />
+        </TestComponentWrapper>
+    );
 
 describe('History', () => {
     let wrapper: ShallowWrapper<History>;
@@ -30,7 +35,7 @@ describe('History', () => {
     });
 
     it('should render', () => {
-        expect(wrapper).toMatchSnapshot();
+        expect(wrapper.render()).toMatchSnapshot();
     });
 
     it('should return correct data', () => {
@@ -48,14 +53,5 @@ describe('History', () => {
 
         const result = mapValues(maxVolume, orderEntry);
         expect(result).toEqual(expectedData);
-    });
-
-    it('should handle onSelect function', () => {
-        // tslint:disable: no-shadowed-variable
-        const wrapper = mount(<OrderBook {...{ ...defaultProps}} />);
-        wrapper.find('tbody').find('tr').first().simulate('click');
-        expect(wrapper.find('tbody').find('tr').first().prop('className')).toEqual('cr-table__row--selected');
-        expect((defaultProps.onSelect as SinonSpy).calledOnceWith()).toBeTruthy();
-        wrapper.unmount();
     });
 });
