@@ -1,9 +1,10 @@
 import * as React from 'react';
-import { Button } from 'react-bootstrap';
 import { captchaLogin } from '../../api';
 import { SignInComponent, SignInProps } from './';
 
 import { shallow } from 'enzyme';
+import { TestComponentWrapper } from 'lib/test';
+import { Button } from 'react-bootstrap';
 
 const defaults: SignInProps = {
     onForgotPassword: jest.fn(),
@@ -30,48 +31,46 @@ const defaults: SignInProps = {
 };
 
 const setup = (props: Partial<SignInProps> = {}) =>
-    shallow(<SignInComponent {...{...defaults, ...props }} />);
-
+    shallow(
+        <TestComponentWrapper>
+            <SignInComponent {...{ ...defaults, ...props }} />
+        </TestComponentWrapper>
+    );
 
 // TODO: We need to rewrite tests in order to test hooks
-describe.skip('SignIn component', () => {
+describe('SignIn component', () => {
     it('should render', () => {
         const wrapper = setup();
-        expect(wrapper).toMatchSnapshot();
-    });
-
-    it('renders without crashing', () => {
-        const wrapper = setup();
-        expect(wrapper).toBeDefined();
+        expect(wrapper.render()).toMatchSnapshot();
     });
 
     it('render correct title', () => {
-        const wrapper = setup();
+        const wrapper = setup().render();
         expect(wrapper.find('.__selected').text()).toBe('Sign In');
     });
 
     it('should render logo block', () => {
-        let wrapper = setup();
+        let wrapper = setup().render();
         const firstState = wrapper.find('.cr-sign-in-form__form-content').children();
         expect(firstState).toHaveLength(4);
-        wrapper = setup({ image: 'image'});
-        const secondState  = wrapper.find('.cr-sign-in-form__form-content').children();
+        wrapper = setup({ image: 'image' }).render();
+        const secondState = wrapper.find('.cr-sign-in-form__form-content').children();
         expect(secondState).toHaveLength(5);
     });
 
     it('should have correct labels', () => {
-        const wrapper = setup({ labelSignIn: 'label sign in', labelSignUp: 'label sign up'});
+        const wrapper = setup({ labelSignIn: 'label sign in', labelSignUp: 'label sign up' }).render();
         expect(wrapper.find('.__selected').text()).toBe('label sign in');
         expect(wrapper.find('.cr-sign-in-form__tab-signup').text()).toBe('label sign up');
     });
 
     it('should render error blocks', () => {
-        const wrapper = setup({emailError: 'error email', passwordError: 'error password'});
+        const wrapper = setup({ emailError: 'error email', passwordError: 'error password' }).render();
         expect(wrapper.find('.cr-sign-in-form__error').first().text()).toBe('error email');
         expect(wrapper.find('.cr-sign-in-form__error').last().text()).toBe('error password');
     });
 
-    it('should send request', () => {
+    it.skip('should send request', () => {
         const spyOnValidateForm = jest.fn();
         const spyOnRefreshError = jest.fn();
         const spyOnSignIn = jest.fn();
@@ -91,7 +90,7 @@ describe.skip('SignIn component', () => {
         expect(spyOnSignIn).toHaveBeenCalledTimes(1);
     });
 
-    it('should validate form', () => {
+    it.skip('should validate form', () => {
         const spyOnValidateForm = jest.fn();
         const spyOnRefreshError = jest.fn();
         const spyOnSignIn = jest.fn();
@@ -111,9 +110,9 @@ describe.skip('SignIn component', () => {
     });
 
     it('should have correct labels for input fields', () => {
-        let wrapper = setup();
+        let wrapper = setup().render();
         expect(wrapper.find('.cr-sign-in-form__bottom-section-password').text()).toBe('Forgot your password?');
-        wrapper = setup({forgotPasswordLabel: 'label forgot password'});
+        wrapper = setup({ forgotPasswordLabel: 'label forgot password' }).render();
         expect(wrapper.find('.cr-sign-in-form__bottom-section-password').text()).toBe('label forgot password');
     });
 
