@@ -1,121 +1,106 @@
 import { mount, shallow } from 'enzyme';
-import * as React from 'react';
-import { WalletItem } from './';
+import { TestComponentWrapper } from 'lib/test';
+import React from 'react';
+import { WalletItem, WalletItemProps } from './';
 
-describe('WalletItem', () => {
-    describe('#render', () => {
-        it('uses styles by default', () => {
-            const wrapper = mount(
-                <WalletItem
-                    currency={'BTC'}
-                    name={'Bitcoin'}
-                    // tslint:disable-next-line:no-magic-numbers
-                    balance={'45'}
-                    // tslint:disable-next-line:no-magic-numbers
-                    locked={'3'}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fee={0.3}
-                    type={'fiat'}
-                    active={false}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fixed={8}
-                />,
-            );
-            const { className } = wrapper.find('div').first().props();
-            expect(className).toContain('cr-wallet-item');
-        });
+const setup = (props: WalletItemProps) =>
+    shallow(
+        <TestComponentWrapper>
+            <WalletItem {...props} />
+        </TestComponentWrapper>
+    );
 
-        it('displays codes, balances and locked', () => {
-            const wrapper = mount(
-                <WalletItem
-                    currency={'BTC'}
-                    name={'Bitcoin'}
-                    balance={'45'}
-                    locked={'3'}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fee={0.3}
-                    type={'fiat'}
-                    active={false}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fixed={8}
-                />,
-            );
-            expect(wrapper.find('.cr-wallet-item__description').text()).toContain('BTCBitcoin');
-        });
+describe('WalletItem #render', () => {
+    it('uses styles by default', () => {
+        const wrapper = setup({
+            address: 'aevrv',
+            currency: 'BTC',
+            name: 'Bitcoin',
+            balance: '45',
+            locked: '3',
+            fee: 0.3,
+            type: 'fiat',
+            active: false,
+            fixed: 8,
+        }).render();
+        expect(wrapper.find('div.cr-wallet-item')).toBeDefined();
+    });
 
-        it('should not display locked balance if wallet does not have one', () => {
-            const wrapper = shallow(
-                <WalletItem
-                    currency={'BTC'}
-                    name={'Bitcoin'}
-                    balance={'45'}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fee={0.3}
-                    type={'fiat'}
-                    active={false}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fixed={8}
-                />,
-            );
-            const balanceElement = wrapper.find('.cr-wallet-item__balance');
-            expect(balanceElement.props().className)
-                .not.toContain('cr-wallet-item__balance-locked');
-        });
+    it('displays codes, balances and locked', () => {
+        const wrapper = setup({
+            address: 'aevrv',
+            currency: 'BTC',
+            name: 'Bitcoin',
+            balance: '45',
+            locked: '3',
+            fee: 0.3,
+            type: 'fiat',
+            active: false,
+            fixed: 8,
+        }).render();
+        expect(wrapper.find('.cr-wallet-item__description').text()).toContain('BTCBitcoin');
+    });
 
-        it('shows locked amount', () => {
-            const wrapper = mount(
-                <WalletItem
-                    currency={'BTC'}
-                    name={'Bitcoin'}
-                    balance={'45'}
-                    locked={'3'}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fee={0.3}
-                    type={'fiat'}
-                    active={true}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fixed={8}
-                />,
-            );
-            const { className } = wrapper.find('div').first().props();
-            expect(className).toContain('cr-wallet-item--active');
-            expect(wrapper.find('.cr-wallet-item__amount-locked').text().trim()).toEqual('3.00000000');
-        });
+    it('should not display locked balance if wallet does not have one', () => {
+        const wrapper = setup({
+            address: 'aevrv',
+            currency: 'BTC',
+            name: 'Bitcoin',
+            balance: '45',
+            locked: '3',
+            fee: 0.3,
+            type: 'fiat',
+            active: false,
+            fixed: 8,
+        }).render();
+        const balanceElement = wrapper.find('.cr-wallet-item__balance');
+        expect(balanceElement.attr('class')).not.toContain('cr-wallet-item__balance-locked');
+    });
 
-        it('should render when locked = 0', () => {
-            const wrapper = mount(
-                <WalletItem
-                    name={'Bitcoin'}
-                    currency={'BTC'}
-                    balance={'45'}
-                    locked={''}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fee={0.3}
-                    type={'fiat'}
-                    active={true}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fixed={8}
-                />,
-            );
-            expect(wrapper.find('.cr-wallet-item__balance-locked').text()).toEqual('');
-        });
+    it('shows locked amount', () => {
+        const wrapper = setup({
+            address: 'aevrv',
+            currency: 'BTC',
+            name: 'Bitcoin',
+            balance: '45',
+            locked: '3',
+            fee: 0.3,
+            type: 'fiat',
+            active: false,
+            fixed: 8,
+        }).render();
+        const className = wrapper.find('div').first().attr('class');
+        expect(className).toContain('cr-wallet-item__info');
+        expect(wrapper.find('.cr-wallet-item__amount-locked').text().trim()).toEqual('3.00000000');
+    });
 
-        it('should match snapshot', () => {
-            const wrapper = mount(
-                <WalletItem
-                    currency={'BTC'}
-                    name={'Bitcoin'}
-                    balance={'45'}
-                    locked={'3'}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fee={0.3}
-                    type={'fiat'}
-                    active={false}
-                    // tslint:disable-next-line:no-magic-numbers
-                    fixed={8}
-                />,
-            );
-            expect(wrapper).toMatchSnapshot();
-        });
+    it('should render when locked = 0', () => {
+        const wrapper = setup({
+            address: 'aevrv',
+            currency: 'BTC',
+            name: 'Bitcoin',
+            balance: '45',
+            locked: '',
+            fee: 0.3,
+            type: 'fiat',
+            active: false,
+            fixed: 8,
+        }).render();
+        expect(wrapper.find('.cr-wallet-item__balance-locked').text()).toEqual('');
+    });
+
+    it('should match snapshot', () => {
+        const wrapper = setup({
+            address: 'aevrv',
+            currency: 'BTC',
+            name: 'Bitcoin',
+            balance: '45',
+            locked: '3',
+            fee: 0.3,
+            type: 'fiat',
+            active: false,
+            fixed: 8,
+        }).render();
+        expect(wrapper).toMatchSnapshot();
     });
 });
