@@ -1,7 +1,4 @@
-import {
-    PG_TITLE_PREFIX,
-    pgRoutes,
-} from './';
+import { PG_TITLE_PREFIX, pgRoutes } from './';
 
 describe('Constants', () => {
     const expectedRoutesForLoggedInUser = [
