@@ -1,105 +1,38 @@
-import classnames from 'classnames';
 import * as React from 'react';
 import { Spinner } from 'react-bootstrap';
-import {
-    injectIntl,
-} from 'react-intl';
-import { connect, MapDispatchToPropsFunction } from 'react-redux';
-import { compose } from 'redux';
-import { IntlProps } from '../../';
+import { useIntl } from 'react-intl';
+import { useDispatch, useSelector } from 'react-redux';
 import { Decimal, Table } from '../../components';
+import { DEFAULT_MARKET } from '../../constants';
 import { localeDate, setTradesType } from '../../helpers';
 import {
     fetchHistory,
-    Market,
-    RootState,
     selectCurrentMarket,
     selectCurrentPrice,
     selectHistory,
     selectHistoryLoading,
     setCurrentPrice,
-    WalletHistoryList,
 } from '../../modules';
 import { handleHighlightValue } from './Market';
 
-interface ReduxProps {
-    list: WalletHistoryList;
-    fetching: boolean;
-    currentMarket: Market | undefined;
-    currentPrice: number | undefined;
-}
-
-interface DispatchProps {
-    fetchHistory: typeof fetchHistory;
-    setCurrentPrice: typeof setCurrentPrice;
-}
-
-type Props = ReduxProps & DispatchProps & IntlProps;
-
 const timeFrom = String(Math.floor((Date.now() - 1000 * 60 * 60 * 24) / 1000));
 
-class RecentTradesYoursContainer extends React.Component<Props> {
+export const RecentTradesYours = () => {
+    const { formatMessage } = useIntl();
+    const dispatch = useDispatch();
 
-    public componentDidMount() {
-        const { currentMarket } = this.props;
-        if (currentMarket) {
-            this.props.fetchHistory({ type: 'trades', page: 0, time_from: timeFrom, market: currentMarket.id} as any);
-        }
-    }
+    const list = useSelector(selectHistory);
+    const fetching = useSelector(selectHistoryLoading);
+    const currentMarket = useSelector(selectCurrentMarket) || DEFAULT_MARKET;
+    const currentPrice = useSelector(selectCurrentPrice);
 
-    public componentWillReceiveProps(next: Props) {
-        if (next.currentMarket && this.props.currentMarket !== next.currentMarket) {
-            this.props.fetchHistory({ type: 'trades', page: 0, time_from: timeFrom, market: next.currentMarket.id });
-        }
-    }
+    const headers = React.useMemo(() => ([
+        formatMessage({ id: 'page.body.trade.header.recentTrades.content.time' }),
+        formatMessage({ id: 'page.body.trade.header.recentTrades.content.amount' }),
+        formatMessage({ id: 'page.body.trade.header.recentTrades.content.price' }),
+    ]), [formatMessage]);
 
-    public shouldComponentUpdate(nextProps: Props) {
-        return JSON.stringify(nextProps.list) !== JSON.stringify(this.props.list);
-    }
-
-    public render() {
-        const { fetching } = this.props;
-        const className = classnames({
-            'cr-tab-content__noData': this.retrieveData()[0][1] === this.props.intl.formatMessage({ id: 'page.noDataToShow' }),
-        });
-
-        return (
-            <div className={className}>
-                {fetching ? <div className="cr-tab-content-loading"><Spinner animation="border" variant="primary" /></div> : this.renderContent()}
-            </div>
-        );
-    }
-
-    public renderContent = () => {
-        return (
-            <React.Fragment>
-                <Table
-                    header={this.getHeaders()}
-                    data={this.retrieveData()}
-                    onSelect={this.handleOnSelect}
-                />
-            </React.Fragment>
-        );
-    };
-
-    private getHeaders = () => {
-        return [
-            this.props.intl.formatMessage({ id: 'page.body.trade.header.recentTrades.content.time' }),
-            this.props.intl.formatMessage({ id: 'page.body.trade.header.recentTrades.content.amount' }),
-            this.props.intl.formatMessage({ id: 'page.body.trade.header.recentTrades.content.price' }),
-        ];
-    };
-
-    private retrieveData = () => {
-        const { list } = this.props;
-
-        return [...list].length > 0
-            ? [...list].map(this.renderRow)
-            : [[[''], this.props.intl.formatMessage({ id: 'page.noDataToShow' })]];
-    };
-
-    private renderRow = (item, i) => {
-        const { currentMarket, list } = this.props;
+    const renderRow = (item, i) => {
         const { id, created_at, price, amount, side } = item;
         const priceFixed = currentMarket ? currentMarket.price_precision : 0;
         const amountFixed = currentMarket ? currentMarket.amount_precision : 0;
@@ -117,30 +50,42 @@ class RecentTradesYoursContainer extends React.Component<Props> {
         ];
     };
 
-    private handleOnSelect = (index: string) => {
-        const { list, currentPrice } = this.props;
+    const retrieveData = () => {
+        return list.length > 0
+            ? list.map(renderRow)
+            : [[]];
+    };
+
+    const renderContent = () => {
+        return (
+            <Table
+                header={headers}
+                data={retrieveData()}
+                onSelect={handleOnSelect}
+            />
+        );
+    };
+
+    const handleOnSelect = (index: string) => {
         const priceToSet = list[Number(index)] ? Number(list[Number(index)].price) : 0;
 
         if (currentPrice !== priceToSet) {
-            this.props.setCurrentPrice(priceToSet);
+            dispatch(setCurrentPrice(priceToSet));
         }
     };
-}
 
-const mapStateToProps = (state: RootState): ReduxProps => ({
-    list: selectHistory(state),
-    fetching: selectHistoryLoading(state),
-    currentMarket: selectCurrentMarket(state),
-    currentPrice: selectCurrentPrice(state),
-});
+    React.useEffect(() => {
+        dispatch(fetchHistory({
+            type: 'trades',
+            page: 0,
+            time_from: timeFrom,
+            market: currentMarket.id,
+        }));
+    }, [dispatch, currentMarket.id]);
 
-const mapDispatchToProps: MapDispatchToPropsFunction<DispatchProps, {}> =
-    dispatch => ({
-        fetchHistory: params => dispatch(fetchHistory(params)),
-        setCurrentPrice: payload => dispatch(setCurrentPrice(payload)),
-    });
-
-export const RecentTradesYours = compose(
-    injectIntl,
-    connect(mapStateToProps, mapDispatchToProps),
-)(RecentTradesYoursContainer) as any; // tslint:disable-line
+    return (
+        <div>
+            {fetching ? <div className="cr-tab-content-loading"><Spinner animation="border" variant="primary" /></div> : renderContent()}
+        </div>
+    );
+};
