@@ -131,13 +131,12 @@ export class WalletTable extends React.Component<Props> {
         const {
             currency,
             currencies,
-            intl,
             type,
             wallets,
         } = this.props;
         const { fixed } = wallets.find(w => w.currency === currency) || { fixed: 8 };
         if (list.length === 0) {
-            return [[intl.formatMessage({ id: 'page.noDataToShow' }), '', '']];
+            return [[]];
         }
 
         return list.sort((a, b) => {
