@@ -2,27 +2,27 @@ import { Currency } from '../modules/public/currencies';
 import * as helpers from './';
 
 describe('Constants regexps', () => {
-  it('should validate correct passwords', () => {
-    expect('1azeRTY@'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
-    expect('1aze@RTY'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
-    expect('azeRTY@2'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
-    expect('azeRTY@222222222'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
-  });
+    it('should validate correct passwords', () => {
+        expect('1azeRTY@'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
+        expect('1aze@RTY'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
+        expect('azeRTY@2'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
+        expect('azeRTY@222222222'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
+    });
 
-  it('should validate correct passwords in non latin languages', () => {
-    expect('йцукRTY@2'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
-    expect('azeЙЦУК@3'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
-  });
+    it('should validate correct passwords in non latin languages', () => {
+        expect('йцукRTY@2'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
+        expect('azeЙЦУК@3'.match(helpers.PASSWORD_REGEX)).toBeTruthy();
+    });
 
-  it('should not validate incorrect passwords', () => {
-    expect(''.match(helpers.PASSWORD_REGEX)).toBeNull();
-    expect('1234568'.match(helpers.PASSWORD_REGEX)).toBeNull();
-    expect('abcdegh'.match(helpers.PASSWORD_REGEX)).toBeNull();
-    expect('ABCDFGH'.match(helpers.PASSWORD_REGEX)).toBeNull();
-    expect('aaaBBBB'.match(helpers.PASSWORD_REGEX)).toBeNull();
-    expect('aZ1@'.match(helpers.PASSWORD_REGEX)).toBeNull();
-    expect('1az@RTY'.match(helpers.PASSWORD_REGEX)).toBeNull();
-  });
+    it('should not validate incorrect passwords', () => {
+        expect(''.match(helpers.PASSWORD_REGEX)).toBeNull();
+        expect('1234568'.match(helpers.PASSWORD_REGEX)).toBeNull();
+        expect('abcdegh'.match(helpers.PASSWORD_REGEX)).toBeNull();
+        expect('ABCDFGH'.match(helpers.PASSWORD_REGEX)).toBeNull();
+        expect('aaaBBBB'.match(helpers.PASSWORD_REGEX)).toBeNull();
+        expect('aZ1@'.match(helpers.PASSWORD_REGEX)).toBeNull();
+        expect('1az@RTY'.match(helpers.PASSWORD_REGEX)).toBeNull();
+    });
 });
 
 describe('Helpers', () => {
@@ -114,7 +114,9 @@ describe('Helpers', () => {
 
     // emailValidation.js
     it('Rendering correct regular expression for email validation', () => {
-        expect(helpers.EMAIL_REGEX).toEqual(/^(?:[\w\!\#\$\%\&\'\*\+\-\/\=\?\^\`\{\|\}\~]+\.)*[\w\!\#\$\%\&\'\*\+\-\/\=\?\^\`\{\|\}\~]+@(?:(?:(?:[a-zA-Z0-9](?:[a-zA-Z0-9\-](?!\.)){0,61}[a-zA-Z0-9]?\.)+[a-zA-Z0-9](?:[a-zA-Z0-9\-](?!$)){0,61}[a-zA-Z0-9]?)|(?:\[(?:(?:[01]?\d{1,2}|2[0-4]\d|25[0-5])\.){3}(?:[01]?\d{1,2}|2[0-4]\d|25[0-5])\]))$/);
+        expect(helpers.EMAIL_REGEX).toEqual(
+            /^(?:[\w\!\#\$\%\&\'\*\+\-\/\=\?\^\`\{\|\}\~]+\.)*[\w\!\#\$\%\&\'\*\+\-\/\=\?\^\`\{\|\}\~]+@(?:(?:(?:[a-zA-Z0-9](?:[a-zA-Z0-9\-](?!\.)){0,61}[a-zA-Z0-9]?\.)+[a-zA-Z0-9](?:[a-zA-Z0-9\-](?!$)){0,61}[a-zA-Z0-9]?)|(?:\[(?:(?:[01]?\d{1,2}|2[0-4]\d|25[0-5])\.){3}(?:[01]?\d{1,2}|2[0-4]\d|25[0-5])\]))$/
+        );
     });
 
     // filterData.js
@@ -129,12 +131,16 @@ describe('Helpers', () => {
     // formatCCYAddress.ts
     it('Should right format BCH address', () => {
         expect(helpers.formatCCYAddress('BTC', '')).toEqual('');
-        expect(helpers.formatCCYAddress('BTC', '2NDJKBjTcoPRXTnkAbcyeqLwXQfdoeug2cL')).toEqual('2NDJKBjTcoPRXTnkAbcyeqLwXQfdoeug2cL');
+        expect(helpers.formatCCYAddress('BTC', '2NDJKBjTcoPRXTnkAbcyeqLwXQfdoeug2cL')).toEqual(
+            '2NDJKBjTcoPRXTnkAbcyeqLwXQfdoeug2cL'
+        );
         expect(helpers.formatCCYAddress('BCH', '')).toEqual('');
         expect(helpers.formatCCYAddress('BCH', '2NDJKBjTcoPRXTnkAbcyeqLwXQfdoeug2c0')).toEqual('');
         expect(helpers.formatCCYAddress('BCH', '2NDJKBjTcoPRXTnkAbcyeqLwXQfdoeug2c')).toEqual('');
         expect(helpers.formatCCYAddress('BCH', '2NDJKBjTcoPRXTnkAbcyeqLwXQfdoeug2cLa')).toEqual('');
-        expect(helpers.formatCCYAddress('BCH', '2NDJKBjTcoPRXTnkAbcyeqLwXQfdoeug2cL')).toEqual('bchtest:prdlsx39pn5v6yxmcgaqgeg6zd48snqaay69chcgwx');
+        expect(helpers.formatCCYAddress('BCH', '2NDJKBjTcoPRXTnkAbcyeqLwXQfdoeug2cL')).toEqual(
+            'bchtest:prdlsx39pn5v6yxmcgaqgeg6zd48snqaay69chcgwx'
+        );
     });
 
     // historyTableUtils.js
@@ -189,7 +195,12 @@ describe('Helpers', () => {
 
     // handleCCYPrecision.ts
     it('Should return correctly precised numbers', () => {
-        const currencies = [{id: 'eth', precision: 5}, {id: 'usd', precision: 4}, {id: 'btc', precision: 8}, {id: 'zar', precision: 8}] as Currency[];
+        const currencies = [
+            { id: 'eth', precision: 5 },
+            { id: 'usd', precision: 4 },
+            { id: 'btc', precision: 8 },
+            { id: 'zar', precision: 8 },
+        ] as Currency[];
         expect(helpers.handleCCYPrecision(currencies, 'btc', 4)).toBe(8);
         expect(helpers.handleCCYPrecision(currencies, 'trst', 4)).toBe(4);
     });
@@ -198,7 +209,7 @@ describe('Helpers', () => {
     it('Should return correct trade color', () => {
         expect(helpers.setTradeColor('buy')).toEqual(tradesColorsMapping.buy);
         expect(helpers.setTradeColor('sell')).toEqual(tradesColorsMapping.sell);
-        expect(helpers.setTradeColor('')).toEqual({ color: ''});
+        expect(helpers.setTradeColor('')).toEqual({ color: '' });
     });
 
     // sliceString.ts
@@ -286,10 +297,10 @@ describe('Helpers', () => {
 
     // getTotalPrice.ts
     it('should return correct total price', () => {
-        expect(helpers.getTotalPrice('0', 10, [['0','0']])).toEqual(0);
-        expect(helpers.getTotalPrice('2', 10, [['0','0']])).toEqual(0);
-        expect(helpers.getTotalPrice('0', 10, [['2','10']])).toEqual(0);
-        expect(helpers.getTotalPrice('2', 10, [['2','10']])).toEqual(4);
+        expect(helpers.getTotalPrice('0', 10, [['0', '0']])).toEqual(0);
+        expect(helpers.getTotalPrice('2', 10, [['0', '0']])).toEqual(0);
+        expect(helpers.getTotalPrice('0', 10, [['2', '10']])).toEqual(0);
+        expect(helpers.getTotalPrice('2', 10, [['2', '10']])).toEqual(4);
         expect(helpers.getTotalPrice('10', 10, proposals)).toEqual(120);
     });
 
@@ -326,24 +337,74 @@ describe('Helpers', () => {
         expect(helpers.getHistorySagaParam({ page: 0, limit: 25, type: 'deposits' })).toBe('page=1&limit=25');
         expect(helpers.getHistorySagaParam({ page: 0, limit: 10, type: 'withdraws' })).toBe('page=1&limit=10');
         expect(helpers.getHistorySagaParam({ page: 0, limit: 5, type: 'trades' })).toBe('page=1&limit=5');
-        expect(helpers.getHistorySagaParam({ page: 0, currency: 'btc', type: 'deposits', limit: 6 })).toBe('page=1&currency=btc&limit=6');
-        expect(helpers.getHistorySagaParam({ page: 0, currency: 'btc', type: 'withdraws', limit: 25 })).toBe('page=1&currency=btc&limit=25');
-        expect(helpers.getHistorySagaParam({ page: 0, limit: 25, type: 'trades', market: 'btcusd' })).toBe('page=1&limit=25&market=btcusd');
-        expect(helpers.getHistorySagaParam({ page: 0, limit: 25, type: 'trades', time_from: '1557878400' })).toBe('page=1&limit=25&time_from=1557878400');
-        expect(helpers.getHistorySagaParam({ page: 0, limit: 25, type: 'trades', time_from: '1557878400', time_to: '1557964799' })).toBe('page=1&limit=25&time_from=1557878400&time_to=1557964799');
-        expect(helpers.getHistorySagaParam({ page: 0, limit: 25, type: 'trades', market: 'btcusd', time_from: '1557878400', time_to: '1557964799' })).toBe('page=1&limit=25&market=btcusd&time_from=1557878400&time_to=1557964799');
-        expect(helpers.getHistorySagaParam({ page: 0, limit: 25, type: 'trades', time_from: '1557878400', time_to: '1557964799' })).toBe('page=1&limit=25&time_from=1557878400&time_to=1557964799');
-
+        expect(helpers.getHistorySagaParam({ page: 0, currency: 'btc', type: 'deposits', limit: 6 })).toBe(
+            'page=1&currency=btc&limit=6'
+        );
+        expect(helpers.getHistorySagaParam({ page: 0, currency: 'btc', type: 'withdraws', limit: 25 })).toBe(
+            'page=1&currency=btc&limit=25'
+        );
+        expect(helpers.getHistorySagaParam({ page: 0, limit: 25, type: 'trades', market: 'btcusd' })).toBe(
+            'page=1&limit=25&market=btcusd'
+        );
+        expect(helpers.getHistorySagaParam({ page: 0, limit: 25, type: 'trades', time_from: '1557878400' })).toBe(
+            'page=1&limit=25&time_from=1557878400'
+        );
+        expect(
+            helpers.getHistorySagaParam({
+                page: 0,
+                limit: 25,
+                type: 'trades',
+                time_from: '1557878400',
+                time_to: '1557964799',
+            })
+        ).toBe('page=1&limit=25&time_from=1557878400&time_to=1557964799');
+        expect(
+            helpers.getHistorySagaParam({
+                page: 0,
+                limit: 25,
+                type: 'trades',
+                market: 'btcusd',
+                time_from: '1557878400',
+                time_to: '1557964799',
+            })
+        ).toBe('page=1&limit=25&market=btcusd&time_from=1557878400&time_to=1557964799');
+        expect(
+            helpers.getHistorySagaParam({
+                page: 0,
+                limit: 25,
+                type: 'trades',
+                time_from: '1557878400',
+                time_to: '1557964799',
+            })
+        ).toBe('page=1&limit=25&time_from=1557878400&time_to=1557964799');
     });
 
     // getUserAgent.ts
     it('Should return correct params', () => {
-        expect(helpers.getUserAgent('Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36')).toBe('Chrome 71 Linux x86_64');
-        expect(helpers.getUserAgent('Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36')).toBe('Chrome 71 Windows 10');
-        expect(helpers.getUserAgent('Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Ubuntu Chromium/73.0.3683.86 Chrome/73.0.3683.86 Safari/537.36')).toBe('Chromium 73 Ubuntu ');
+        expect(
+            helpers.getUserAgent(
+                'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36'
+            )
+        ).toBe('Chrome 71 Linux x86_64');
+        expect(
+            helpers.getUserAgent(
+                'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36'
+            )
+        ).toBe('Chrome 71 Windows 10');
+        expect(
+            helpers.getUserAgent(
+                'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Ubuntu Chromium/73.0.3683.86 Chrome/73.0.3683.86 Safari/537.36'
+            )
+        ).toBe('Chromium 73 Ubuntu ');
         expect(helpers.getUserAgent('PostmanRuntime/7.6.0')).toBe('PostmanRuntime/7.6.0');
-        expect(helpers.getUserAgent('Mozilla/5.0 (Macintosh; U; PPC Mac OS X 10.4; en; rv:1.9.0.19) Gecko/2011091218 Camino/2.0.9 (like Firefox/3.0.19)')).toBe('Camino 2 Mac OS 10.4');
-        expect(helpers.getUserAgent('Mozilla/4.0 (compatible; MSIE 5.0; Windows 95) Opera 6.01 [en]')).toBe('Opera 6 Windows 95');
+        expect(
+            helpers.getUserAgent(
+                'Mozilla/5.0 (Macintosh; U; PPC Mac OS X 10.4; en; rv:1.9.0.19) Gecko/2011091218 Camino/2.0.9 (like Firefox/3.0.19)'
+            )
+        ).toBe('Camino 2 Mac OS 10.4');
+        expect(helpers.getUserAgent('Mozilla/4.0 (compatible; MSIE 5.0; Windows 95) Opera 6.01 [en]')).toBe(
+            'Opera 6 Windows 95'
+        );
     });
 
     // checkDate.ts
@@ -392,6 +453,16 @@ describe('Helpers', () => {
             expect(helpers.truncateMiddle('return correct value', 10)).toBe('retu...lue');
             expect(helpers.truncateMiddle('return correct value', 30)).toBe('return correct value');
             expect(helpers.truncateMiddle('return correct value', 10, '****')).toBe('ret****lue');
+        });
+    });
+
+    describe('hasDuplicates', () => {
+        it('should return true if a value is in array', () => {
+            expect(helpers.hasDuplicates([1, 2, 3], 2)).toBeTruthy();
+        });
+
+        it('should return false if a value is not in array', () => {
+            expect(helpers.hasDuplicates([1, 2, 3], 5)).toBeFalsy();
         });
     });
 });
