@@ -10,7 +10,8 @@ export * from './emailValidation';
 export * from './filterData';
 export * from './formatCCYAddress';
 export * from './formatDate';
-export * from './geetest';
+// MUST BE REMOVED: Vadym P. responsibility
+// export * from './geetest';
 export * from './getCsrfToken';
 export * from './getHistorySagaParam';
 export * from './getLanguageByCode';
@@ -40,3 +41,7 @@ export * from './timezone';
 export * from './toggleColorTheme';
 export * from './truncateMiddle';
 export * from './nicknameValidation';
+export * from './randomSecretHex';
+export * from './getVerificationStep';
+export * from './convertToString';
+export * from './hasDuplicates';
